+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
+## [0.18.0] - 2025-05-22
+MR #20
+
+### Added
+- Alembic migration for the new field `name` in transport table
+- Extraction and insertion of `name` for transport table
+- Dashboard visualisations
+- Queries for the dashboard
+
 ## [0.17.0] - 2025-05-18
 MR #19
 
@@ -14,6 +24,7 @@ MR #19
 - Test route
 - API readme part
 - Heathcheck for database docker compose config
+
 
 ## [0.16.0] - 2025-05-15
 MR #17